ya se pueden agregar las imágenes de los diseños en fotmato jpg, png o gif, pero ya que para visualizar la imagen se debe hacer uso del campo nombre, todas las imagenes deben ser, o solo jpg, o solo png, se arreglaron detalles de corrección de inicio de sesion y cositas varias pensientes, como que ahora a las variantes tambien se les almacena el campo de id de categoria para que sean mostrados todos los diseños en el live preview

+++ sis.info/eliminar.php
@@ -1,16 +1,30 @@
-<?PHP
-/*Requiere la conexion a la base de datos sisvose*/
-include ("funcion/conectarse.php");
-$id=$_GET['id'];
-/*Se eliminan los datos de usuario segun el id*/
+<?php
+include("funcion/conectarse.php");
+require_once("sesion.class.php");
 
-$query="DELETE FROM user WHERE codigo='$id'";
-$resultado=$conexion->query($query);
+$sesion = new sesion();
+$usuario = $sesion->get("usuario");
+
+if( $usuario == false )
+{
+    header("Location: cerrarsesion.php");
+}
+else {
 
 
-if ($resultado>0){
+    ?>
+    <?PHP
+    include("funcion/conectarse.php");
+    $id = $_GET['id'];
+    /*Se eliminan los datos de usuario segun el id*/
 
-    echo "
+    $query = "DELETE FROM user WHERE codigo='$id'";
+    $resultado = $conexion->query($query);
+
+
+    if ($resultado > 0) {
+
+        echo "
     <html>
     <head>
     <meta http-equiv='refresh' content='0; url=users.php'>
@@ -19,10 +33,9 @@ if ($resultado>0){
     </script>
     </head>
     </html>";
-}
-else{
+    } else {
 
-    echo "
+        echo "
     <html>
     <head>
     <meta http-equiv='refresh' content='0; url=users.php'>
@@ -31,9 +44,9 @@ else{
     </script>
     </head>
     </html>";
+    }
+
 }
-
-
 ?>
